refactor(event cache): move handling of a backpagination in `RoomEventCacheState`
